backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
supabase
pydantic
python-dotenv
